+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Apply MatCap color alpha to converted Toon Standard materials.
+- Apply MatCap main color power to converted Toon Standard materials.
+- Apply rim lighting color alpha to converted Toon Standard materials.
+
+### Fixed
+- Aspect ratio not kept when converting textures for Toon Standard.
+- Unexpected shadow ramp when converting materials which don't use shadow to Toon Standard.
+
 ## [2.10.0-beta.2] - 2025-05-20
 
 ### Fixed